requirements.txt
django<3,>=2
psycopg2==2.7.3.2
dj-database-url==0.4.2
python-decouple==3.1
django-modeltranslation==0.12.2
django-debug-toolbar==1.9.1
django-extensions==1.9.9
dj-database-url==0.4.2
Pillow==5.0.0
psycopg2==2.7.3.2
dj-database-url==0.4.2
python-decouple==3.1
django-modeltranslation==0.12.2
django-debug-toolbar==1.9.1
django-extensions==1.9.9
dj-database-url==0.4.2
Pillow==5.0.0
django-tables2
celery[redis]==4.1.0
django-solo==1.1.3
django_addview